# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scandere AI Store API",
        "description": "REST API for digital products e-commerce platform. Sell and manage digital documents (PDFs, Excel, Word) with secure downloads, payments via Stripe, and comprehensive admin panel.",
        "contact": {
            "name": "Scandere Support",
            "email": "team@scandere.info"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.scandereai.store",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "Admin - Orders"
                ],
                "summary": "List all orders (admin)",
                "description": "Get paginated list of all orders with user and product information, plus total revenue and order count. Requires admin authentication.",
                "operationId": "f954f3dfa0d14a0dbea3336954f7bace",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders list with statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orders": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "total_revenue": {
                                            "description": "Total revenue from completed orders",
                                            "type": "number",
                                            "format": "float",
                                            "example": 12345.67
                                        },
                                        "total_orders": {
                                            "description": "Total number of completed orders",
                                            "type": "integer",
                                            "example": 248
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
}